refactor proposal workflow decision routes
