+++ src/app/routers/proposal_workflow.py
@@ -1,7 +1,6 @@
 from fastapi import APIRouter, Header, Path
 
 from app.contracts.proposals import (
-    ProposalApprovalActionRequest,
     ProposalStateTransitionEnvelopeResponse,
     ProposalSubmitRequest,
 )
@@ -38,102 +37,6 @@ async def submit_proposal(
         expected_state=request.expected_state,
         review_type=request.review_type,
         reason=request.reason,
-        related_version_no=request.related_version_no,
-        idempotency_key=idempotency_key,
-        correlation_id=correlation_id,
-    )
-
-
-@router.post(
-    "/{proposal_id}/approve-risk",
-    response_model=ProposalStateTransitionEnvelopeResponse,
-    summary="Approve Proposal Risk Review",
-    description="Records the risk approval decision for a proposal in risk review.",
-)
-async def approve_risk(
-    request: ProposalApprovalActionRequest,
-    proposal_id: str = Path(
-        ...,
-        description="Gateway-visible proposal identifier returned by lotus-advise.",
-        examples=["pp_1"],
-    ),
-    idempotency_key: str = Header(
-        alias="Idempotency-Key",
-        description="Caller-supplied idempotency key for proposal risk-approval requests.",
-        examples=["idem-approve-risk-1"],
-    ),
-) -> ProposalStateTransitionEnvelopeResponse:
-    service = proposal_service()
-    correlation_id = correlation_id_var.get()
-    return await service.approve_risk(
-        proposal_id=proposal_id,
-        actor_id=request.actor_id,
-        expected_state=request.expected_state,
-        details=request.details,
-        related_version_no=request.related_version_no,
-        idempotency_key=idempotency_key,
-        correlation_id=correlation_id,
-    )
-
-
-@router.post(
-    "/{proposal_id}/approve-compliance",
-    response_model=ProposalStateTransitionEnvelopeResponse,
-    summary="Approve Proposal Compliance Review",
-    description="Records the compliance approval decision for a proposal in compliance review.",
-)
-async def approve_compliance(
-    request: ProposalApprovalActionRequest,
-    proposal_id: str = Path(
-        ...,
-        description="Gateway-visible proposal identifier returned by lotus-advise.",
-        examples=["pp_1"],
-    ),
-    idempotency_key: str = Header(
-        alias="Idempotency-Key",
-        description="Caller-supplied idempotency key for proposal compliance-approval requests.",
-        examples=["idem-approve-compliance-1"],
-    ),
-) -> ProposalStateTransitionEnvelopeResponse:
-    service = proposal_service()
-    correlation_id = correlation_id_var.get()
-    return await service.approve_compliance(
-        proposal_id=proposal_id,
-        actor_id=request.actor_id,
-        expected_state=request.expected_state,
-        details=request.details,
-        related_version_no=request.related_version_no,
-        idempotency_key=idempotency_key,
-        correlation_id=correlation_id,
-    )
-
-
-@router.post(
-    "/{proposal_id}/record-client-consent",
-    response_model=ProposalStateTransitionEnvelopeResponse,
-    summary="Record Proposal Client Consent",
-    description="Records client consent for a proposal that has completed internal approvals.",
-)
-async def record_client_consent(
-    request: ProposalApprovalActionRequest,
-    proposal_id: str = Path(
-        ...,
-        description="Gateway-visible proposal identifier returned by lotus-advise.",
-        examples=["pp_1"],
-    ),
-    idempotency_key: str = Header(
-        alias="Idempotency-Key",
-        description="Caller-supplied idempotency key for client-consent recording requests.",
-        examples=["idem-client-consent-1"],
-    ),
-) -> ProposalStateTransitionEnvelopeResponse:
-    service = proposal_service()
-    correlation_id = correlation_id_var.get()
-    return await service.record_client_consent(
-        proposal_id=proposal_id,
-        actor_id=request.actor_id,
-        expected_state=request.expected_state,
-        details=request.details,
         related_version_no=request.related_version_no,
         idempotency_key=idempotency_key,
         correlation_id=correlation_id,
